Fix incorrect PLAY without pids
When the PLAY command continues using an already opened SAT>IP session, then the `pids=` parameter is required.
This fixes the problem when before a TEARDOWN command a new internal session is created. Without this patch a `pids=none` is included in the PLAY command and the session never starts.