Encapsulate HTTP responses when using persistent requests
When using the HTTP protocol for streaming, it's possible to process multiple requests (aka Persistent Mode). And this is great! However, when doing the streaming the HTTP response is sended in between of the TS packets. Then it's possible to "encapsulate" the response inside a NULL TS packet.

This patch implements this and allows the client to change pids or request new streams without interrupting the streaming.