Fix no TEARDOWN with satipc module (v.2)
This patch implements a new optional adapter callback function named standby(). This function is called when the adapter doesn't have any stream clients running. This function is previous to the close() adapter. The expected behaviour is to put the tuner in a standby sleep for a fast restart in the future.

This new functionality is used in the satipc module. With the new functionality when the last client sends a TEARDOWN command, the module sends a TEARDOWN to the remote SAT>IP server, but it doesn't close the connection. Then any new tuning command from a SAT>IP client can be served instantly. In any case, when the timeout of the adapter is reached, then the adapter is closed definitively.

This PR supersedes catalinii#728.