satipc: don't send a TEARDOWN without stream
In some very specific cases, for example when the SAT>IP server doesn't sends any RTP data, the device can be closed without a valid stream id. In this case, it's not required to send the TEARDOWN message. Futhermore, some SAT>IP servers have toubles when receiving a TEARDOWN message without the stream parameter, what actually is a break of the specifications (the TEARDOWN command **requires** the `/stream=` part in the URI).

This fixes this problem.