FIX of debug -l socketworks
When enabling the debug of the "socketworks" module the enabled module is the "socket" one, becase the name "socket" is contained inside "socketworks".
This fixes the problem enabling only a module when the exact name matches. To do this is used a new function maps_intd() that is used only here.