Fix missing PCR pid (#816)
When the PCR of a service is independent (i.e., it has no associated stream) the PCR pid is not added to the list of the pids. This patch fixes this problem. So when simulating the pids=all in E2 boxes the PCR pid is added to the list of pids.