Enhanced drop of non-decrypted packets (#819)
* Limit the wait time to 40ms

When the calculated max wait time (4 * UDP threshold) is more than the PCR interval, several clients will be in troubles.
So limit the wait time to 40ms at maximum.

* Enhanced drop of non-decrypted packets

When a packet can't be decrypted (and the drop encrypted packets is enabled) instead of filter out all encrypted packets do a more smart strategy. The implemented strategy consists of passing only packets with PCR time stamps. The passed PCR packets will don't have any payload data, so only the TS header is preserved. This will be more friendly for decoders that require the PCR marks. Futhermore the client will continue receiving packets, and then the stream can be considered active. Without this patch when a decoding disruption is a bit long several clients will close the session if it is not repaired in a short time.

* Rollback 0x1FFF check

* Safe clean TSC even if no Payload is found

Even if this case has no sense, clean the bits if for some reason TSC (encrypted) bits are set when no payload is carried.