Update satipc_read() with recvmmsg() (#827)
* Update satipc_read() with recvmmsg()

When using the satipc module the reading of the RTP packets requires to be as fast as you can (without overloading the process).
Until now the UDP reading thread was reading only one packet at a time. So when receiving a chunk of packets it will be preferable to read as many as can be put into the read buffer. This new implementation uses the function `recvmmsg()` to just use one syscall to read more than one packet in one shot. The maximum number of packets to read is defined by `MAX_RTP_MSG`. The current value is sufficient to not generate oversubscription and to avoid delaying the packets too much when the bitrate rises a lot at some point in time.

* Fix the implementation of the satipc_read() 

When using the recvmmsg() syscall the holes are identified and processed after (only if they exist).
Futhermore, support for an emulated recvmmsg() is included for platforms without it.