Search for slave adapters
This fixes the problem when searching for free adapters. Before this patch no adapters are found when all master tuners are in use and slave adapters are free. This patch solves this problem searching over the slave adapters after the initial search. Now the slave adapters will be used when needed.