Use the new "name_app" option
Added a new "name_app" option to change the announced name of the server. By default the "minisatip" is the name of the server, but with this new option it could be set to any name. This is useful to use more than one servers in the same LAN and distinguish them. The optional name is used only for advertaising, so the device continues to be "minisatip". Useful for example when using several instances with clients like VideoLAN.