Calculate SNR dB to Percentage based on Modulation
This patch incorporates a SNR value conversion from dB scale to percentage scale.
To do it, it uses a table of the correct dB values for each modulation. These values correspond to a 100% good signal.
The table includes at time only the common modulations. Other values could be included in the future.