Send diseqc command with 18V
When multiple tuners share the same wire in a master-slave configuration, it's preferable to send all diseqc commands using the 18V voltage. Then commands are more easy to read from devices when tuners inject incorrect power or slave tuners are miss configured. This behaviour increase the robustness of the diseqc commands.