Implementation of smooth unencrypted packets
Now when "drop_encrypted == 2" instead of filter out all non-PCR packets, we convert all unencrypted packets to TS packets with stuffing removing the PES payload. This will maintain the number of TS packets and the CC counters. Therefore, the resulting bitrate is identical to that of the original encrypted stream. And that it's more easy to handle for regular hardware and software decoders. Producing then a more smooth play when some part of the stream can't be decrypted.