Add Tuner status type enforcement
For some adapters the DVB API version 5.5 is not available. In these cases the software can't distinguish the status reporting of the tuners. Therefore, the values of Strength and/or SNR could be incorrect. If you know how the driver of the tuner reports the data (in percentage or in dB), then you could tweak the values with the `-M --multiplier` parameter.

This patch adds the option to enforce percentage or decibel modes at tuner level (because you can mix different hardware and drivers in the same machine). Example:

`-M 0:#1.0-@1.0 -M 1:#1.0-@1.0 -M 2:2.0-1.0 -M 3:2.0-1.0`

In this case the first two tuners (DVB-S) use default values enforcing "percentage" for Strength and "dB" for SNR. And for the other two tuners (DVB-T2) the types aren't enforced but the Strength is doubled.

This removes all dependencies for Enigma2 boxes, and you will have the opportunity to set the correct types/values for all of your tuners.