Hide HTTP response of a reGET (#1031)
When one HTTP client wants to change the pid list it could send a new GET command identical to the first one with different pids.
The current sever code can handle this non-standard use of the HTTP protocol. But without this extension the HTTP response is introduced in the middle of the streaming. And this generates noise from the client point of view.
This new implementation changes the behaviour checking if the client sends multiple GET commands. In this case the response for the successive requests is automatically hidden. The result is that the streaming continues from the server to the client without any interruption.