Restore socket buffer configuration (#1038)
* Restore socket buffer configuration

This patch restores the lost option to set the socket buffers. However, now it's possible to set independently the read socket buffer for the satipc connections, and the write socket buffer for the clients. By default the buffer uses the kernel configuration. However, with two new parameters the user can overwrite them.

* fix compilation in some environments

* Fix stupid typo of name of statement