Fix compare tunning if SR=auto in current tunning (#1037)
When comparing new and current tunning values, the SR could be AUTO. Not only for the "new" tunning, but also for the "current" tunning.
This fixes the tunning when the "current" adapter tuning is using AUTO mode for SR, and the "new" request is forcing some value. In this edge case, we can consider that the current running value is the one corresponding to the current tunning mode. That's we trust the new request.