Send all ECMs optionally (#1035)
* Resend ECMs optionally

By default when a key is invalid the same ECM is not resended to the server. This default behaviour has sense when the server has returned an invalid key for the ECM. However, with some busy cards resending the same ECM to the server will obtain a valid key. A new option is added to handle this specific case.

* Fix compilation error with some compilers

* Use more simple syntax

* Update opts.h

* Update minisatip.c

* Update dvbapi.c