Use a vector of descriptors to store stream PID descriptors
Previously we stored all descriptors as raw data in a single pointer. This is inflexible since we can't easily access individual descriptors if we want to. We also had to allocate a large chunk of memory for the descriptors to make sure we were able to fit all of them.

The next step would be to use the same mechanism for storing PMT descriptors too, instead of what we do now where we parse CA descriptors and store the CAIDs and CA PIDs separately. Since we can now loop through descriptors, we no longer need to store information from them separately.